+++ R/MPN.R
@@ -19,27 +19,24 @@
 #' Haas et al. (2014).
 #'
 #' @references Bacteriological Analytical Manual, 8th Edition, Appendix 2,
-#'   \url{https://www.fda.gov/Food/FoodScienceResearch/LaboratoryMethods/ucm109656.htm}
+#'   \url{https://www.fda.gov/food/laboratory-methods-food/bam-appendix-2-most-probable-number-serial-dilutions}
 #'
 #' @references Bacteriological Analytical Manual, 8th Edition, Chapter 3,
-#'   \url{https://www.fda.gov/food/foodscienceresearch/laboratorymethods/ucm063346.htm}
+#'   \url{https://www.fda.gov/food/laboratory-methods-food/bam-chapter-3-aerobic-plate-count}
 #'
 #' @references Blodgett RJ (2002). "Measuring improbability of outcomes from a
 #'   serial dilution test." \emph{Communications in Statistics: Theory and
 #'   Methods}, 31(12), 2209-2223.
-#'   \url{https://www.tandfonline.com/doi/abs/10.1081/STA-120017222}
 #'
 #' @references Blodgett RJ (2005). "Serial dilution with a confirmation step."
 #'   \emph{Food Microbiology}, 22(6), 547-552.
-#'   \url{https://doi.org/10.1016/j.fm.2004.11.017}
 #'
 #' @references Blodgett RJ (2010). "Does a serial dilution experiment's model
 #'   agree with its outcome?" \emph{Model Assisted Statistics and Applications},
-#'   5(3), 209-215. \url{https://doi.org/10.3233/MAS-2010-0157}
+#'   5(3), 209-215.
 #'
 #' @references Haas CN, Heller B (1988). "Averaging of TNTC counts."
 #'   \emph{Applied and Environmental Microbiology}, 54(8), 2069-2072.
-#'   \url{https://aem.asm.org/content/54/8/2069}
 #'
 #' @references Haas CN (1989). "Estimation of microbial densities from dilution
 #'   count experiments" \emph{Applied and Environmental Microbiology} 55(8),
@@ -52,7 +49,7 @@
 #' @references Jarvis B, Wilrich C, Wilrich P-T (2010). "Reconsideration of the
 #'   derivation of Most Probable Numbers, their standard deviations, confidence
 #'   bounds and rarity values." \emph{Journal of Applied Microbiology}, 109,
-#'   1660-1667. \url{https://doi.org/10.1111/j.1365-2672.2010.04792.x}
+#'   1660-1667.
 #'
 #' @references Ridout MS (1994). "A comparison of confidence interval methods
 #'   for dilution series experiments." \emph{Biometrics}, 50(1), 289-296.
@@ -60,7 +57,7 @@
 #' @references Salama IA, Koch GG, Tolley DH. (1978) "On the estimation of the
 #'   most probable number in a serial dilution technique." \emph{Communications
 #'   in Statistics - Theory and Methods}, 7(13), 1267-1281.
-#'
-#' @docType package
-#' @name MPN
+#' @keywords internal
+"_PACKAGE"
+
 NULL